Change vocodec channels to have 2 output interfaces
After the input is decoded, it is copied to two internal voice queues - one for each output interface. The 3xxx interface thread task has been updated and a new interface implementation has been added for Codec 2.

For 3xxx interfaces, when the device returns an encoded voice packet, the interface channel number is no longer enough to find the transcoding channel, as the same interface with the same interface channel will now appear into two transcoding channel configurations of any given triplet. To solve this, the current open transcoding channel for each interface channel has been added as state in the respective interface. Still need to protect this with some lock.

Also missing is the part that creates the transcoding triplets and mark the whole group as unavailable when one of the transcoding channels is in use.